chore(tests): move helpers and fixtures to new directories
