+++ tests/__fixtures__/errorPackage/package.json
