+++ tests/__fixtures__/errorPackage/src/index.js
