+++ tests/__fixtures__/exportsPackage/package.json
