+++ tests/__fixtures__/exportsPackage/src/index.js
