+++ tests/__fixtures__/externalDepPackage/package.json
