+++ tests/__fixtures__/externalDepPackage/src/index.js
