+++ tests/__fixtures__/jsonPackage/package.json
