+++ tests/__fixtures__/jsonPackage/src/index.js
