+++ tests/__fixtures__/testPackage/src/index.js
