+++ tests/__helpers__/bundleChecks.js
