+++ tests/__helpers__/createDummyStream.js
