+++ tests/__helpers__/removeArtifacts.js
